split out the db routines and merges the user.js files
