+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "fs-app-template",
+  "name": "team2coffee",
   "version": "2.0.0",
   "description": "Some boilerplate code to get you started - get shakin'!",
   "main": "index.js",
