+++ server/db/db.js
@@ -11,6 +11,8 @@ if(process.env.LOGGING === 'true'){
   delete config.logging
 }
 
+const dbid = process.env.DBIDREQUIRED;
+
 //https://stackoverflow.com/questions/61254851/heroku-postgres-sequelize-no-pg-hba-conf-entry-for-host
 if(process.env.DATABASE_URL){
   config.dialectOptions = {
@@ -21,5 +23,6 @@ if(process.env.DATABASE_URL){
 }
 
 const db = new Sequelize(
-  process.env.DATABASE_URL || `postgres://localhost:5432/${databaseName}`, config)
+  process.env.DATABASE_URL || `postgres://${dbid}localhost:5432/${databaseName}`, config)
+
 module.exports = db
